+++ lib/services/audio_input_ffi_stub.dart
@@ -313,6 +313,8 @@ class AudioInputFFI {
   int get rehFramesPerBeat => 0;
   int get rehFramesPerBar => 0;
   int rehAddTrack(String wavPath) => -1;
+
+  int stretchFile(String inPath, String outPath, double ratio) => -1;
   void rehRemoveTrack(int idx) {}
   void rehClearTracks() {}
   void rehSetTrackGain(int idx, double gain) {}